+++ meteor-app/imports/startup/server/methods-payments.js
@@ -3,6 +3,7 @@ import axios from 'axios'
 import { Carts } from '/imports/api/products/schema'
 import CONSTANTS from '/imports/api/constants'
 import Members from '/imports/api/members/schema'
+import Charges from '/imports/api/charges/schema'
 
 const debug = require('debug')('b2b:payments')
 
@@ -177,6 +178,39 @@ Meteor.methods({
       debug(error)
       // throw new Meteor.Error(error.message)
       return error.message
+    }
+  },
+  'refresh.charges': async function() {
+    const chargeURL = `${Meteor.settings.private.paymentURL}`
+    const request = {
+      url: chargeURL,
+      method: 'get',
+      validateStatus: function(status) {
+        return status >= 200
+      },
+      auth: {
+        username: Meteor.settings.private.paymentApiKey,
+        password: ''
+      }
+    }
+    debug(`Fetching charges from ${chargeURL}`, request)
+    try {
+      let n = 0
+      const r = await axios(request)
+      const { data } = r
+      debug(`status: ${r.status} ${r.statusText}`, data.response)
+      data.response.forEach(charge => {
+        const c = Charges.findOne({ token: charge.token })
+        if (!c) {
+          const id = Charges.insert(charge)
+          if (id) n = n + 1
+        }
+      })
+      return { status: 'success', message: `Added ${n} charges` }
+    } catch (error) {
+      debug(error)
+      // throw new Meteor.Error(error.message)
+      return { status: 'failed', message: error.message }
     }
   }
 })