+++ meteor-app/imports/ui/service/service-item-search.js
@@ -21,7 +21,6 @@ const SearchBar = props => {
       })}
       results={results}
       value={value}
-      resultRenderer={resultRenderer}
     />
   )
 }